old code cleanup
