+++ public/components/create_job/job_stages/index_selection/index_selection.js
