+++ public/components/create_job/job_stages/mapping_selection/example_queries.json
