+++ public/components/create_job/job_stages/mapping_selection/mapping_field.js
