+++ public/components/health_check/healthcheck.js
