+++ public/components/health_check/kibana.html
