+++ public/components/list_jobs/job_entry.js
@@ -5,7 +5,7 @@ import PropTypes from 'prop-types';
 
 import moment from 'moment';
 
-import { GenericRequest } from '../../utils/requests';
+import { GenericRequest } from '~services';
 
 import {
