+++ public/components/list_jobs/job_header.js
@@ -33,7 +33,8 @@ export class JobListHeader extends React.Component {
   }
 
   onClickCreateJob = () => {
-    window.location.href = '#/create_job';
+    this.props.history.push('create_job');
+    // this.componentWillUnmount();
   }
