+++ public/components/main.js
