+++ public/components/mainContent.js
