--- public/components/v2/create_job/job_stages/mapping_selection/example_queries.json
@@ -1,39 +1,0 @@
-[
-  {
-    "query": {
-        "match_all": {}
-    }
-  },
-  {
-    "query": {
-      "bool": {
-        "filter": {
-          "term": {
-            "status": "active"
-          }
-        }
-      }
-    }
-  },
-  {
-    "query": { 
-      "bool": { 
-        "must": [
-          { "match": { "title":   "Search"        }}, 
-          { "match": { "content": "Elasticsearch" }}  
-        ],
-        "filter": [ 
-          { "term":  { "status": "published" }}, 
-          { "range": { "publish_date": { "gte": "2015-01-01" }}} 
-        ]
-      }
-    }
-  },
-  {
-    "query": {
-        "match" : {
-            "message" : "this is a test"
-        }
-    }
-  }
-]
