--- public/components/v2/health_check/kibana.html
@@ -1,2 +1,0 @@
-arcanna.endpoint.url: "http://<HOST>:<PORT>"
-arcanna.endpoint.token: "<AUTHENTICATION_TOKEN>"
